Added web module with url operations.
Updated get_url_domain name to get_url_hostname.
Updated get_url_hostname to return the hostname.
Renamed get_url to build_url.
Added web module to setup.py.
Quality of code improvements.
Removed cyclic imports.
Added line separation on lines greater then 100 characters.
