+++ yk_utils/objects/object_operations.py
@@ -1,6 +1,4 @@
 """ Objects operations file """
-from urllib.parse import urlparse, urljoin
-from yk_utils.objects.exceptions import BadUrl
 import datetime
 
 
@@ -23,36 +21,3 @@ def datetime2str(value: datetime.datetime):
         raise TypeError("Value is not datetime.")
 
     return value.__str__()
-
-
-def get_url_domain(url: str) -> str:
-    """
-    Gets the domain of the given url
-    :param url: url
-    :return: url main domain
-    """
-    return urlparse(url).hostname.split('.')[-2]
-
-
-def is_valid_url(url: str) -> bool:
-    """
-    Checks the format validity of the given url.
-    :param url: url
-    :return: True if valid. False otherwise.
-    """
-    result = urlparse(url)
-    return all([result.scheme, result.netloc])
-
-
-def get_url(base: str, relative: str = "") -> str:
-    """
-    Joins both base and relative url and checks for its validity.
-    If not valid will raise BadUrl exception.
-    :param base: base url path to be specified
-    :param relative: relative path to be specified
-    :return:
-    """
-    absolute = urljoin(base, relative)
-    if is_valid_url(absolute):
-        return absolute
-    raise BadUrl(f"Bad URL '{absolute}'")
